pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "langchain-chroma>=1.1.0",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "typing-extensions>=4.15.0",
]
